http2: Abstracting stream and session-related logic
