+++ lib/internal/http2/util.js
@@ -12,6 +12,11 @@ const {
   Symbol,
 } = primordials;
 
+const {
+  kUpdateTimer,
+  kHandle,
+} = require('internal/stream_base_commons');
+
 const binding = internalBinding('http2');
 const {
   codes: {
@@ -20,7 +25,8 @@ const {
     ERR_HTTP2_INVALID_PSEUDOHEADER,
     ERR_HTTP2_INVALID_SETTING_VALUE,
     ERR_INVALID_ARG_TYPE,
-    ERR_INVALID_HTTP_TOKEN
+    ERR_INVALID_HTTP_TOKEN,
+    ERR_INVALID_OPT_VALUE,
   },
   addCodeToName,
   hideStackFrames
@@ -28,12 +34,14 @@ const {
 
 const kSensitiveHeaders = Symbol('nodejs.http2.sensitiveHeaders');
 const kSocket = Symbol('socket');
+const kState = Symbol('state');
 const kProxySocket = Symbol('proxySocket');
 const kRequest = Symbol('request');
 
 const {
   NGHTTP2_NV_FLAG_NONE,
   NGHTTP2_NV_FLAG_NO_INDEX,
+  NGHTTP2_DEFAULT_WEIGHT,
   NGHTTP2_SESSION_CLIENT,
   NGHTTP2_SESSION_SERVER,
 
@@ -210,6 +218,10 @@ const IDX_OPTIONS_MAX_OUTSTANDING_SETTINGS = 7;
 const IDX_OPTIONS_MAX_SESSION_MEMORY = 8;
 const IDX_OPTIONS_MAX_SETTINGS = 9;
 const IDX_OPTIONS_FLAGS = 10;
+
+const kMaxFrameSize = (2 ** 24) - 1;
+const kMaxInt = (2 ** 32) - 1;
+const kMaxStreams = (2 ** 32) - 1;
 
 function updateOptionsBuffer(options) {
   let flags = 0;
@@ -614,25 +626,127 @@ function sessionName(type) {
   }
 }
 
+// Validates the values in a settings object. Specifically:
+// 1. headerTableSize must be a number in the range 0 <= n <= kMaxInt
+// 2. initialWindowSize must be a number in the range 0 <= n <= kMaxInt
+// 3. maxFrameSize must be a number in the range 16384 <= n <= kMaxFrameSize
+// 4. maxConcurrentStreams must be a number in the range 0 <= n <= kMaxStreams
+// 5. maxHeaderListSize must be a number in the range 0 <= n <= kMaxInt
+// 6. enablePush must be a boolean
+// 7. enableConnectProtocol must be a boolean
+// All settings are optional and may be left undefined
+const validateSettings = hideStackFrames((settings) => {
+  if (settings === undefined) return;
+  assertWithinRange('headerTableSize',
+                    settings.headerTableSize,
+                    0, kMaxInt);
+  assertWithinRange('initialWindowSize',
+                    settings.initialWindowSize,
+                    0, kMaxInt);
+  assertWithinRange('maxFrameSize',
+                    settings.maxFrameSize,
+                    16384, kMaxFrameSize);
+  assertWithinRange('maxConcurrentStreams',
+                    settings.maxConcurrentStreams,
+                    0, kMaxStreams);
+  assertWithinRange('maxHeaderListSize',
+                    settings.maxHeaderListSize,
+                    0, kMaxInt);
+  if (settings.enablePush !== undefined &&
+      typeof settings.enablePush !== 'boolean') {
+    throw new ERR_HTTP2_INVALID_SETTING_VALUE('enablePush',
+                                              settings.enablePush);
+  }
+  if (settings.enableConnectProtocol !== undefined &&
+      typeof settings.enableConnectProtocol !== 'boolean') {
+    throw new ERR_HTTP2_INVALID_SETTING_VALUE('enableConnectProtocol',
+                                              settings.enableConnectProtocol);
+  }
+});
+
+// Validates that priority options are correct, specifically:
+// 1. options.weight must be a number
+// 2. options.parent must be a positive number
+// 3. options.exclusive must be a boolean
+// 4. if specified, options.silent must be a boolean
+//
+// Also sets the default priority options if they are not set.
+const setAndValidatePriorityOptions = hideStackFrames((options) => {
+  if (options.weight === undefined) {
+    options.weight = NGHTTP2_DEFAULT_WEIGHT;
+  } else if (typeof options.weight !== 'number') {
+    throw new ERR_INVALID_OPT_VALUE('weight', options.weight);
+  }
+
+  if (options.parent === undefined) {
+    options.parent = 0;
+  } else if (typeof options.parent !== 'number' || options.parent < 0) {
+    throw new ERR_INVALID_OPT_VALUE('parent', options.parent);
+  }
+
+  if (options.exclusive === undefined) {
+    options.exclusive = false;
+  } else if (typeof options.exclusive !== 'boolean') {
+    throw new ERR_INVALID_OPT_VALUE('exclusive', options.exclusive);
+  }
+
+  if (options.silent === undefined) {
+    options.silent = false;
+  } else if (typeof options.silent !== 'boolean') {
+    throw new ERR_INVALID_OPT_VALUE('silent', options.silent);
+  }
+});
+
+function callTimeout(self, kSession) {
+  // If the session is destroyed, this should never actually be invoked,
+  // but just in case...
+  if (self.destroyed)
+    return;
+  // This checks whether a write is currently in progress and also whether
+  // that write is actually sending data across the write. The kHandle
+  // stored `chunksSentSinceLastWrite` is only updated when a timeout event
+  // happens, meaning that if a write is ongoing it should never equal the
+  // newly fetched, updated value.
+  if (self[kState].writeQueueSize > 0) {
+    const handle = kSession ? self[kSession][kHandle] : self[kHandle];
+    const chunksSentSinceLastWrite = handle !== undefined ?
+      handle.chunksSentSinceLastWrite : null;
+    if (chunksSentSinceLastWrite !== null &&
+      chunksSentSinceLastWrite !== handle.updateChunksSent()) {
+      self[kUpdateTimer]();
+      return;
+    }
+  }
+
+  self.emit('timeout');
+}
+
 module.exports = {
   assertIsObject,
   assertValidPseudoHeader,
   assertValidPseudoHeaderResponse,
   assertValidPseudoHeaderTrailer,
   assertWithinRange,
+  callTimeout,
   getDefaultSettings,
   getSessionState,
   getSettings,
   getStreamState,
   isPayloadMeaningless,
   kSensitiveHeaders,
+  kMaxFrameSize,
+  kMaxInt,
+  kMaxStreams,
   kSocket,
+  kState,
   kProxySocket,
   kRequest,
   mapToHeaders,
   NghttpError,
   sessionName,
+  setAndValidatePriorityOptions,
   toHeaderObject,
   updateOptionsBuffer,
-  updateSettingsBuffer
+  updateSettingsBuffer,
+  validateSettings,
 };
