+++ node.gyp
@@ -219,6 +219,8 @@
       'lib/internal/util/types.js',
       'lib/internal/http2/core.js',
       'lib/internal/http2/compat.js',
+      'lib/internal/http2/session.js',
+      'lib/internal/http2/streams.js',
       'lib/internal/http2/util.js',
       'lib/internal/v8_prof_polyfill.js',
       'lib/internal/v8_prof_processor.js',
